[SCB-2722]governance add a Mapper that give a map result if request match
