+++ governance/src/main/java/org/apache/servicecomb/governance/handler/FaultInjectionHandler.java
@@ -20,8 +20,8 @@ package org.apache.servicecomb.governance.handler;
 import org.apache.servicecomb.governance.marker.GovernanceRequest;
 import org.apache.servicecomb.governance.policy.FaultInjectionPolicy;
 import org.apache.servicecomb.governance.properties.FaultInjectionProperties;
-import org.apache.servicecomb.injection.Fault;
-import org.apache.servicecomb.injection.FaultInjectionUtil;
+import org.apache.servicecomb.governance.processor.injection.Fault;
+import org.apache.servicecomb.governance.processor.injection.FaultInjectionUtil;
 
 public class FaultInjectionHandler extends AbstractGovernanceHandler<Fault, FaultInjectionPolicy> {
