+++ governance/src/main/java/org/apache/servicecomb/governance/policy/MapperPolicy.java
@@ -15,31 +15,18 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.injection;
+package org.apache.servicecomb.governance.policy;
 
-import io.vavr.CheckedFunction0;
+import java.util.Map;
 
-public interface FaultInjectionDecorators {
-  static <T> FaultInjectionDecorateCheckedSupplier<T> ofCheckedSupplier(CheckedFunction0<T> supplier) {
-    return new FaultInjectionDecorateCheckedSupplier<>(supplier);
+public class MapperPolicy extends AbstractPolicy {
+  private Map<String, String> target;
+
+  public Map<String, String> getTarget() {
+    return target;
   }
 
-  class FaultInjectionDecorateCheckedSupplier<T> {
-
-    private CheckedFunction0<T> supplier;
-
-    protected FaultInjectionDecorateCheckedSupplier(CheckedFunction0<T> supplier) {
-      this.supplier = supplier;
-    }
-
-    public FaultInjectionDecorateCheckedSupplier<T> withFaultInjection(Fault fault) {
-      supplier = Fault.decorateCheckedSupplier(fault, supplier);
-      return this;
-    }
-
-    public T get() throws Throwable {
-      return supplier.apply();
-    }
+  public void setTarget(Map<String, String> target) {
+    this.target = target;
   }
-
 }
