+++ governance/src/main/java/org/apache/servicecomb/governance/processor/injection/FaultInjectionException.java
@@ -15,12 +15,19 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.injection;
+package org.apache.servicecomb.governance.processor.injection;
 
-public interface Sleepable {
-  /**
-   * sleep some time
-   * @param delay time unit is millisecond
-   */
-  void sleep(long delay);
+public class FaultInjectionException extends RuntimeException {
+  private static final long serialVersionUID = 1675558351029273343L;
+
+  private final FaultResponse faultResponse;
+
+  public FaultInjectionException(FaultResponse faultResponse) {
+    super(faultResponse.getErrorMsg());
+    this.faultResponse = faultResponse;
+  }
+
+  public FaultResponse getFaultResponse() {
+    return faultResponse;
+  }
 }
