+++ governance/src/main/java/org/apache/servicecomb/governance/processor/loadbanlance/LoadBalance.java
@@ -15,40 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.servicecomb.injection;
+package org.apache.servicecomb.governance.processor.loadbanlance;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.servicecomb.governance.policy.LoadBalancerPolicy;
 
-/**
- * Fault injection parameters which decides the fault injection condition.
- */
-public class FaultParam {
-  private static final Logger LOGGER = LoggerFactory.getLogger(FaultParam.class);
+public interface LoadBalance {
 
-  private final long reqCount;
-
-  private Sleepable sleepable = (delay) -> {
-    try {
-      Thread.sleep(delay);
-    } catch (InterruptedException e) {
-      LOGGER.info("Interrupted exception is received");
-    }
-  };
-
-  public long getReqCount() {
-    return reqCount;
+  static LoadBalance getLoadBalance(String key, LoadBalancerPolicy policy) {
+    LoadBalance loadBalance = new LoadBalanceImpl(policy.getRule());
+    return loadBalance;
   }
 
-  public FaultParam(long reqCount) {
-    this.reqCount = reqCount;
-  }
-
-  public Sleepable getSleepable() {
-    return sleepable;
-  }
-
-  public void setSleepable(Sleepable sleepable) {
-    this.sleepable = sleepable;
-  }
+  String getRule();
 }
