+++ governance/src/main/java/org/apache/servicecomb/governance/properties/MapperProperties.java
@@ -1,0 +1,37 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.servicecomb.governance.properties;
+
+import org.apache.servicecomb.governance.policy.MapperPolicy;
+
+public class MapperProperties extends PolicyProperties<MapperPolicy> {
+  private static final String MATCH_MAPPER_KEY = "servicecomb.mapper";
+
+  public MapperProperties() {
+    super(MATCH_MAPPER_KEY);
+  }
+
+  public MapperProperties(String key) {
+    super(key);
+  }
+
+  @Override
+  protected Class<MapperPolicy> getEntityClass() {
+    return MapperPolicy.class;
+  }
+}
